HW-1 adapter와 Holder 클래스 분리
